% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatelm.R
\name{estatelm}
\alias{estatelm}
\title{Fitting estate avaliation Linear Models}
\usage{
estatelm(
  formula,
  data,
  subset = NULL,
  transf = NULL,
  mode = c("rural", "urban"),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that
can be coerced to that class): a symbolic description of
the model to be fitted..}

\item{y}{an optional data frame, list or environment (or object
coercible by as.data.frame to a data frame) containing the
variables in the model. If not found in data, the variables are
taken from environment(formula), typically the environment from
which lm is called.}
}
\value{
estatelm return an object of class "estatelm".
}
\description{
Fitting estate avaliation Linear Models
}
\examples{
data(exemplo)
estatelm(preco ~ area + quartos + banheiros, data = exemplo)
}
