% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAvaliation.R
\name{readAvaliation}
\alias{readAvaliation}
\title{Identifies the arquive type and read}
\usage{
readAvaliation(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}
}
\value{
estatelm return an object of class "estate.data.frame".
}
\description{
Identifies the arquive type and read
}
\examples{
data_csv <- readAvaliation(file = "~/pesquisa.csv")
}
